+++ src/dbally/__init__.py
@@ -8,10 +8,11 @@ from dbally.views.sqlalchemy_base import SqlAlchemyBaseView
 from dbally.views.structured import BaseStructuredView
 
 from .__version__ import __version__
-from ._exceptions import DbAllyError, LLMConnectionError, LLMError, LLMResponseError, LLMStatusError
+from ._exceptions import DbAllyError
 from ._main import create_collection
 from ._types import NOT_GIVEN, NotGiven
 from .collection import Collection
+from .llms.clients._exceptions import LLMConnectionError, LLMError, LLMResponseError, LLMStatusError
 
 __all__ = [
     "__version__",
