+++ src/dbally/llms/clients/litellm.py
@@ -8,8 +8,8 @@ from dbally.data_models.audit import LLMEvent
 from dbally.llms.clients.base import LLMClient, LLMOptions
 from dbally.prompts import ChatFormat
 
-from ..._exceptions import LLMConnectionError, LLMResponseError, LLMStatusError
 from ..._types import NOT_GIVEN, NotGiven
+from ._exceptions import LLMConnectionError, LLMResponseError, LLMStatusError
 
 
 @dataclass
